+++ Assets/Azure/Resource/ResourceObjectRootObject.cs
@@ -1,13 +1,15 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using Newtonsoft.Json;
 
 namespace Assets.Azure.Resource
 {
     [Serializable]
     public class ResourceObjectRootObject
     {
-        public List<ResourceObject> value { get; set; }
+        [JsonProperty(PropertyName = "value")]
+        public List<ResourceObject> Value { get; set; }
     }
 }