+++ Assets/DataCenter/Gnome/Scripts/HideArea.cs
@@ -1,18 +1,18 @@
-﻿using System.Collections;
+using System.Collections;
 using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
 public class HideArea : MonoBehaviour {
 
-    List<HideParent> hideParents { get; set; }
+    List<HideParent> HideParents { get; set; }
 
 	// Use this for initialization
 	void Start () {
-        hideParents = new List<HideParent>();
-        hideParents.AddRange(transform.parent.GetComponentsInChildren<HideParent>());
-        hideParents.AddRange(transform.parent.GetComponents<HideParent>());
-        foreach (var item in hideParents)
+        HideParents = new List<HideParent>();
+        HideParents.AddRange(transform.parent.GetComponentsInChildren<HideParent>());
+        HideParents.AddRange(transform.parent.GetComponents<HideParent>());
+        foreach (var item in HideParents)
         {
             item.OnHideDestroyed += DestroyHide; 
         }