+++ Assets/Scripts/AzureManagementAPI.cs
@@ -1,4 +1,4 @@
-﻿using Assets.Azure.Subscription;
+using Assets.Azure.Subscription;
 using Newtonsoft.Json;
 using System;
 using System.Collections;
@@ -16,19 +16,19 @@ namespace Assets.Scripts
         public List<Subscription> Subscriptions;
         public delegate void SubscriptionLoadedAction(List<Subscription> loadedSubscriptions);
         public static event SubscriptionLoadedAction OnSubscriptionLoaded;
-        public IEnumerator GetSubscriptions(string UserAccessToken)
+        public IEnumerator GetSubscriptions(string userAccessToken)
         {
             var apiUrl = "https://management.azure.com/subscriptions?api-version=2014-04-01";
 
-            UnityWebRequest Request = UnityWebRequest.Get(apiUrl);
+            UnityWebRequest request = UnityWebRequest.Get(apiUrl);
 
-            Request.method = UnityWebRequest.kHttpVerbGET;
-            Request.SetRequestHeader("Content-Type", "application/json; utf-8");
-            Request.SetRequestHeader("Authorization", UserAccessToken);
+            request.method = UnityWebRequest.kHttpVerbGET;
+            request.SetRequestHeader("Content-Type", "application/json; utf-8");
+            request.SetRequestHeader("Authorization", userAccessToken);
 
-            yield return Request.SendWebRequest();
-            var JsonString = Request.downloadHandler.text;
-            Subscriptions = JsonConvert.DeserializeObject<SubscriptionRootObject>(JsonString).value;
+            yield return request.SendWebRequest();
+            var jsonString = request.downloadHandler.text;
+            Subscriptions = JsonConvert.DeserializeObject<SubscriptionRootObject>(jsonString).Value;
             OnSubscriptionLoaded(Subscriptions);
         }
     }