+++ ChangeLog.md
@@ -1,3 +1,7 @@
+## Develop
+* Code refactoring
+* Added help text to login screen
+
 ## 0.1.4
 * Removed dev servers